Use mmap to allocate cells and mprotect to watch them
Cells are allocated using mmap() to be able to easily mprotect() cell data
to detect mutator reads/writes to freed/swept cells. Using mprotect()
makes the program die with a SIGSEGV *exactly* where it tried to use a
free cell.

This patch assumes a page size of 4096, it uses a full page for the cell
header to be able to mprotect just the cell data (because mprotect ca only
protect full pages), allowing the GC to keep manipulating cell headers as
needed.